Added navigation to API documentation, renamed route
