+++ app/client/src/styles.css
@@ -157,7 +157,3 @@
 .swagger-ui .info {
   margin-top: 0;
 }
-
-hgroup.main a.link {
-  display: none;
-}
